backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
psycopg2-binary
asyncpg
